Show 150 weekly status entries on page
By default 25 entries per page are shown, which includes the RSS feed.
Because the feed doesn't support pagination on a monday if you fetch the
feed in the evening some entries are missing. This commit introduces a
workaround by showing 150 entries per page. I tried to add pagination to
the feed but kaminari (the gem used for pagination) doesn't support
generating paginating links for ATOM feeds...